create subscription

+++ services/service-subscription-service/src/main/java/com/sritel/service_subscription_service/controller/ServiceSubscriptionController.java
@@ -16,7 +16,7 @@ public class ServiceSubscriptionController {
     private final JwtService jwtService;
 
     @PostMapping("/activate/{serviceId}")
-    public ResponseEntity<String> activateService(@PathVariable("serviceId") Long serviceId) {
+    public ResponseEntity<String> activateService(@PathVariable("serviceId") int serviceId) {
         serviceSubscriptionService.activateService(jwtService.extractUserId(), serviceId);
         return ResponseEntity.ok("Service activated successfully");
     }
